refactor: rename app layer to service, add golangci-lint config and comprehensive tests
Rename `internal/app/` to `internal/service/` to better reflect the application
service layer in hexagonal architecture. Add `.golangci.yml` with depguard rules
to enforce layer boundaries. Add comprehensive table-driven tests for all auth
and health service methods. Remove stale mocks, update JWT type definitions,
and fix various middleware/infrastructure issues.
