+++ internal/adapter/outbound/cache/redis/token/token.go
@@ -1,8 +1,6 @@
 package token
 
 import (
-	"go-simple-template/internal/core/port/outbound/cache"
-
 	red "github.com/redis/go-redis/v9"
 )
 
@@ -10,7 +8,7 @@ type cacheRepo struct {
 	client *red.Client
 }
 
-func NewCache(client *red.Client) cache.Token {
+func NewCache(client *red.Client) *cacheRepo {
 	return &cacheRepo{
 		client: client,
 	}
