+++ internal/adapter/outbound/repository/mongo/health/health.go
@@ -1,8 +1,6 @@
 package health
 
 import (
-	"go-simple-template/internal/core/port/outbound/repository"
-
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
@@ -10,7 +8,7 @@ type repo struct {
 	client *mongo.Client
 }
 
-func New(client *mongo.Client) repository.Health {
+func New(client *mongo.Client) *repo {
 	return &repo{
 		client: client,
 	}
