+++ internal/core/port/outbound/cache/mocks/token_mock.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -package mocks -source=token.go -destination=mocks/token_mock.go Token
+//	mockgen -package mocks -source=token.go -destination=mocks/token_mock.go TokenBlacklist
 //
 
 // Package mocks is a generated GoMock package.
@@ -17,32 +17,32 @@ import (
 	gomock "go.uber.org/mock/gomock"
 )
 
-// MockToken is a mock of Token interface.
-type MockToken struct {
+// MockTokenBlacklist is a mock of TokenBlacklist interface.
+type MockTokenBlacklist struct {
 	ctrl     *gomock.Controller
-	recorder *MockTokenMockRecorder
+	recorder *MockTokenBlacklistMockRecorder
 	isgomock struct{}
 }
 
-// MockTokenMockRecorder is the mock recorder for MockToken.
-type MockTokenMockRecorder struct {
-	mock *MockToken
+// MockTokenBlacklistMockRecorder is the mock recorder for MockTokenBlacklist.
+type MockTokenBlacklistMockRecorder struct {
+	mock *MockTokenBlacklist
 }
 
-// NewMockToken creates a new mock instance.
-func NewMockToken(ctrl *gomock.Controller) *MockToken {
-	mock := &MockToken{ctrl: ctrl}
-	mock.recorder = &MockTokenMockRecorder{mock}
+// NewMockTokenBlacklist creates a new mock instance.
+func NewMockTokenBlacklist(ctrl *gomock.Controller) *MockTokenBlacklist {
+	mock := &MockTokenBlacklist{ctrl: ctrl}
+	mock.recorder = &MockTokenBlacklistMockRecorder{mock}
 	return mock
 }
 
 // EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockToken) EXPECT() *MockTokenMockRecorder {
+func (m *MockTokenBlacklist) EXPECT() *MockTokenBlacklistMockRecorder {
 	return m.recorder
 }
 
 // Blacklist mocks base method.
-func (m *MockToken) Blacklist(ctx context.Context, token string, expiration time.Duration) error {
+func (m *MockTokenBlacklist) Blacklist(ctx context.Context, token string, expiration time.Duration) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Blacklist", ctx, token, expiration)
 	ret0, _ := ret[0].(error)
@@ -50,13 +50,13 @@ func (m *MockToken) Blacklist(ctx context.Context, token string, expiration time
 }
 
 // Blacklist indicates an expected call of Blacklist.
-func (mr *MockTokenMockRecorder) Blacklist(ctx, token, expiration any) *gomock.Call {
+func (mr *MockTokenBlacklistMockRecorder) Blacklist(ctx, token, expiration any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Blacklist", reflect.TypeOf((*MockToken)(nil).Blacklist), ctx, token, expiration)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Blacklist", reflect.TypeOf((*MockTokenBlacklist)(nil).Blacklist), ctx, token, expiration)
 }
 
 // IsBlacklisted mocks base method.
-func (m *MockToken) IsBlacklisted(ctx context.Context, token string) (bool, error) {
+func (m *MockTokenBlacklist) IsBlacklisted(ctx context.Context, token string) (bool, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "IsBlacklisted", ctx, token)
 	ret0, _ := ret[0].(bool)
@@ -65,7 +65,7 @@ func (m *MockToken) IsBlacklisted(ctx context.Context, token string) (bool, erro
 }
 
 // IsBlacklisted indicates an expected call of IsBlacklisted.
-func (mr *MockTokenMockRecorder) IsBlacklisted(ctx, token any) *gomock.Call {
+func (mr *MockTokenBlacklistMockRecorder) IsBlacklisted(ctx, token any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsBlacklisted", reflect.TypeOf((*MockToken)(nil).IsBlacklisted), ctx, token)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsBlacklisted", reflect.TypeOf((*MockTokenBlacklist)(nil).IsBlacklisted), ctx, token)
 }
