+++ internal/core/port/outbound/cache/token.go
@@ -5,8 +5,8 @@ import (
 	"time"
 )
 
-//go:generate mockgen -package mocks -source=token.go -destination=mocks/token_mock.go Token
-type Token interface {
+//go:generate mockgen -package mocks -source=token.go -destination=mocks/token_mock.go TokenBlacklist
+type TokenBlacklist interface {
 	Blacklist(ctx context.Context, token string, expiration time.Duration) error
 	IsBlacklisted(ctx context.Context, token string) (bool, error)
 }
