+++ internal/core/port/outbound/repository/mocks/user_mock.go
@@ -70,18 +70,3 @@ func (mr *MockUserRepositoryMockRecorder) Insert(ctx, request any) *gomock.Call 
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Insert", reflect.TypeOf((*MockUserRepository)(nil).Insert), ctx, request)
 }
-
-// Update mocks base method.
-func (m *MockUserRepository) Update(ctx context.Context, request entity.User) (*entity.User, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Update", ctx, request)
-	ret0, _ := ret[0].(*entity.User)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// Update indicates an expected call of Update.
-func (mr *MockUserRepositoryMockRecorder) Update(ctx, request any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockUserRepository)(nil).Update), ctx, request)
-}
