+++ internal/service/auth/auth.go
@@ -1,20 +1,19 @@
 package auth
 
 import (
-	"go-simple-template/internal/core/port/inbound"
 	"go-simple-template/internal/core/port/outbound/cache"
 	"go-simple-template/internal/core/port/outbound/repository"
 )
 
 type auth struct {
 	userRepo   repository.UserRepository
-	tokenCache cache.Token
+	tokenCache cache.TokenBlacklist
 }
 
 func New(
 	userRepo repository.UserRepository,
-	tokenCache cache.Token,
-) inbound.AuthService {
+	tokenCache cache.TokenBlacklist,
+) *auth {
 	return &auth{
 		userRepo:   userRepo,
 		tokenCache: tokenCache,
