+++ internal/service/auth/register.go
@@ -16,14 +16,11 @@ func (s *auth) Register(ctx context.Context, request entity.User) (*entity.User,
 		return nil, err
 	}
 
-	// check if user already exists
 	if user != nil {
 		slog.ErrorContext(ctx, "User already exists", slog.String("email", request.Email))
-
 		return nil, errpkg.NewConflict(errpkg.ErrMsgEmailAlreadyExists)
 	}
 
-	// create hash for password
 	hashedPassword, err := crypto.HashPassword(request.Password)
 	if err != nil {
 		slog.ErrorContext(ctx, "Failed to hash password", slog.String("error", err.Error()))
@@ -32,7 +29,6 @@ func (s *auth) Register(ctx context.Context, request entity.User) (*entity.User,
 
 	request.Password = hashedPassword
 
-	// create user
 	response, err := s.userRepo.Insert(ctx, request)
 	if err != nil {
 		slog.ErrorContext(ctx, "Failed to create user", slog.String("error", err.Error()))
